Refactor property prediction filtering
